Prevent accidental mouse selection of code line numbers
People often copy&paste code/text from code boxes with a mouse.
In the former CSS the selection of line numbers happened too often
because the line numbers had CSS padding on the right side.
By moving this padding to the left side of the text we can better prevent
an accidental selection of line numbers.