fork-awesome style tag should not be conditional anymore
Since now the fork-awesome icons are being used for the mobile navigation toggle buttion, the inclusion of the fork-awesome should not be conditional. If not, the icon won't appear on mobile if you don't configure any links on the theme config, which, as far I as know, is not a required config item.